Refactor: Move streaming functions to dedicated task modules
