+++ backend/src/main.py
@@ -1,4 +1,3 @@
-import asyncio
 import os
 import pathlib
 from concurrent.futures import ThreadPoolExecutor
@@ -6,20 +5,16 @@ from concurrent.futures import ThreadPoolExecutor
 import dotenv
 import uvicorn
 
+from tasks.notification import start_websocket_client
+from tasks.scheduler import fetch_scheduler
 from server import app
 import utils
-from parallel import stream_notification, fetch_scheduler
 from database import Base, engine
 from log import setup_logging
 
 
 def start_api_server():
-    port = int(os.environ.get("PORT"))
-    uvicorn.run(app, host='0.0.0.0', port=port, log_level='info', access_log=False)
-
-
-def start_websocket_client():
-    asyncio.run(stream_notification(os.environ.get("NOTIFICATION_SERVER_URL")))
+    uvicorn.run(app, host='0.0.0.0', port=int(os.environ.get("PORT")), log_level='info', access_log=False)
 
 
 def main():
@@ -35,7 +30,7 @@ def main():
         executor.submit(start_api_server)
         if os.environ.get("AUTO_RECORDING").lower() in ('true', 'enable', 'on'):
             executor.submit(start_websocket_client)
-        executor.submit(fetch_scheduler)
+        executor.submit(fetch_scheduler, os.environ.get("PORT"))
 
 
 if __name__ == "__main__":
